TL: moved RegisterParams and ReadOnlyError classes in proper modules
